requirements.txt
Flask
mysql-connector
python-dateutil
flask-restful
xmltodict
openvas_lib
